refactor: reorganize questionnaire store following modular architecture
This commit refactors the large monolithic questionnaire store (1600+ lines) into a modular architecture:

- Split functionality into domain-specific modules:
  - state.js: Isolates initial state definition
  - getters.js: Contains all computed properties
  - actions/core.js: Core operations (initialize, save, load)
  - actions/questions.js: Question-related operations
  - actions/sections.js: Section management
  - actions/ui.js: UI state management
  - actions/publishing.js: Export and publish functionality
  - helpers/utils.js: Extracted utility functions

- Added comprehensive JSDoc documentation to all modules
- Maintained backward compatibility through a wrapper module (questionnaire.js)
- Improved code organization following SOLID principles
- Reduced module coupling for better testability and maintenance
- No functional changes or breaking changes introduced

This reorganization makes the codebase more maintainable, easier to understand,
and facilitates future enhancements by providing clear separation of concerns.
